SCB-1368 Remove persistent queues for reliability
