+++ alpha/alpha-fsm/src/main/java/org/apache/servicecomb/pack/alpha/fsm/repository/channel/DefaultTransactionRepositoryChannel.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.servicecomb.pack.alpha.fsm.repository.channel;
+
+import org.apache.servicecomb.pack.alpha.core.fsm.repository.model.GlobalTransaction;
+import org.apache.servicecomb.pack.alpha.fsm.metrics.MetricsService;
+import org.apache.servicecomb.pack.alpha.fsm.repository.AbstractTransactionRepositoryChannel;
+import org.apache.servicecomb.pack.alpha.fsm.repository.TransactionRepository;
+
+public class DefaultTransactionRepositoryChannel extends AbstractTransactionRepositoryChannel {
+
+  public DefaultTransactionRepositoryChannel(TransactionRepository repository, MetricsService metricsService) {
+    super(repository, metricsService);
+  }
+
+  @Override
+  public void sendTo(GlobalTransaction transaction) {
+    try {
+      repository.send(transaction);
+    } catch (Exception e) {
+      throw new RuntimeException(e);
+    }
+  }
+}
